requirements.txt
faiss_cpu==1.9.0.post1
huggingface_hub==0.27.0
pandas==2.2.3
sentence_transformers==3.3.1
streamlit==1.41.1
torch==2.5.1
transformers==4.47.1
